// src/app/core/models/ms-order-service/order-payment.model.ts
export interface IOrderPayment {
  id: number;
  orderPayDate: string;
  areaApplicant: string;
  addressAreaApplication: string;
  beneficiary: string;
  contractNumber: string;
  operationType: string;
  description: string;
  awardType: string;
  conditionsPay: string;
  instructionsspecials: string;
  nameelaborate: string;
  userCreation: string;
  userModification: string;
  creationDate: string;
  modificationDate: string;
  version: string;
  instancebpm: string;
  delegationRegionalId: number;
  status: string;
  orderIncomeId: number;
  Subtotal: number;
  amountvat: number;
  total: number;
  applicationpreorderPayId: number;
  confirmationPaysirsaeId: number;
  rfcTea: string;
  reasonSocial: string;
  applicationpreorderDate: string;
  zoneKey: string;
  billSupplierNumber: number;
  billSupplierDate: string;
  documentationannexed: string;
  vouchersNumber: string;
  shapePay: string;
  statussiab: string;
  transfereeId: number;
  nbOrigin: string;
  delegationName: string;
  transferentName: string;
}